// src/app/model/car.model.ts
import {Agence} from './agence.model';


export class CarModel {
    _id: string;
  num: string;
  mark: string;
  model: string;
  state: string;

  // state:string;

  serialNum: string;
  dateFirstRegistration: Date;
  exploitationCartDate: Date;
  exploitationCartNum: string;
    agency: Agence;


    
    


}


 